+++ src/main/kotlin/io/github/ackeecz/danger/lint/FileFinder.kt
@@ -8,14 +8,12 @@ import kotlin.io.path.name
 internal object FileFinder {
 
     private const val BUILD_DIR_NAME = "build"
-    private const val REPORT_FILE_EXTENSION = "xml"
 
-    fun findFiles(
+    fun findBuildDirs(
         rootDirectoryPath: Path,
         buildFoldersMatcher: BuildFoldersMatcher,
-        reportFilesFolderPath: String,
     ): List<File> {
-        val buildDirs = when (buildFoldersMatcher) {
+        return when (buildFoldersMatcher) {
             is BuildFoldersMatcher.All -> {
                 val maxDepth = 6
                 Files.find(
@@ -27,14 +25,38 @@ internal object FileFinder {
                 ).map { it.toFile() }.toList()
             }
             is BuildFoldersMatcher.Specific -> {
-                buildFoldersMatcher.paths.map { File(rootDirectoryPath.toFile(), it) }
+                val resolvedRoot = rootDirectoryPath.toAbsolutePath().toFile()
+                buildFoldersMatcher.paths.map { File(resolvedRoot, it) }
             }
         }
+    }
+
+    fun findFiles(
+        rootDirectoryPath: Path,
+        buildFoldersMatcher: BuildFoldersMatcher,
+        reportFilesFolderPath: String,
+        filePrefix: String = "",
+        fileExtension: String = "xml",
+    ): List<File> {
+        return findFiles(
+            buildDirs = findBuildDirs(rootDirectoryPath, buildFoldersMatcher),
+            reportFilesFolderPath = reportFilesFolderPath,
+            filePrefix = filePrefix,
+            fileExtension = fileExtension,
+        )
+    }
+
+    fun findFiles(
+        buildDirs: List<File>,
+        reportFilesFolderPath: String,
+        filePrefix: String = "",
+        fileExtension: String = "xml",
+    ): List<File> {
         return buildDirs
             .flatMap { buildDir ->
                 File(buildDir, reportFilesFolderPath)
                     .listFiles()
-                    ?.filter { it.extension == REPORT_FILE_EXTENSION }
+                    ?.filter { it.extension == fileExtension && it.nameWithoutExtension.startsWith(filePrefix) }
                     .orEmpty()
             }.ifEmpty {
                 throw NoFilesFoundException()